Day-5-capstone-kaustubh/requirements.txt
fastapi
uvicorn
pandas
requests
pytest
python-dotenv